webkit-efl/LayoutTests/platform/efl-slp/media/video-seek-by-small-increment-expected.txt

CONSOLE MESSAGE: line 27: TypeError: 'undefined' is not a function (evaluating 'element.canPlayType(codecs[i][0])')
FAIL: Timed out waiting for notifyDone to be called
Test seeking by very small increments.
